split tests out to test_sequence_validation
